Fix spike detection for low spikes
The threshold for detecting spike end was set too low, and would never be
reached during the injection window. As a result, just one "spike" would
be detected encompassing a large part of the injection window. The
solution appears to be the make the threshold higher, i.e. easier to pass.