Add spike_ahp_position feature and completely rework ahp_curve_fitness
Previous approach did not work well when there were no spikes:
the score would be lower (better) for that case than for spikes
present but badly shaped. Let's instead make sure that the score
is always higher when an expected spike is missing, and gets lower
when spikes are closes to the expected shape.

Also, values are now constrained in [0, 1] range. This makes it
easier to use this fitness function in new_combined_fitness.

The 20-percentile-best spike is used for comparison. The assumption
is that if some AHPs have good shape, then most likely the shape
is generally correct. The best-fitting AHP might be an outlier, hence
the maneuver with percentiles.