features: split baseline (optionally) into baseline_pre and _post
For some parameter values and higher injection currents we
never return to the baseline, so the "baseline" after injection
period is much higher. Instead of having an in-between value with
very high uncertainty, look at the pre and post parts separately.

This might help by making the baseline more heavily weighted in fitness
functions too.