fitnesses: include spike_count_fitness in new_combined_fitness
spike_time_fitness does not work well if there are no spikes:
it gives a very good score, even though it's not correct to
say that the spikes don't match. It's not entirely correct to
say that they don't match, so we're in a pickle. Adding the count
should alleviate this problem somewhat.