Make save_params faster based on start and threshold.
Before this change, save_params was very slow, as the fitness was
recalculated for every evaluation
whether it met the start or fit thresholds.
Now, only evaluates the fitnesses for
evaluations that meet specified thresholds.
Additionally, thresholds are applied to save npz and save sas,
and an option for saving sas and or npz is added (defaults to True for both).